+++ examples/foxes/multi_height/run.py
@@ -1,5 +1,6 @@
 import time
 import argparse
+import matplotlib.pyplot as plt
 
 import foxes
 import foxes.variables as FV
@@ -14,23 +15,24 @@ def run_foxes(args):
     ttype = foxes.models.turbine_types.PCtFile(args.turbine_file)
     mbook.turbine_types[ttype.name] = ttype
 
-    states = foxes.input.states.StatesTable(
+    states = foxes.input.states.MultiHeightTimeseries(
         data_source=args.states,
-        output_vars=[FV.WS, FV.WD, FV.TI, FV.RHO, FV.MOL],
-        var2col={FV.WS: "ws", FV.WD: "wd", FV.TI: "ti", FV.MOL: "mol"},
-        fixed_vars={FV.RHO: 1.225, FV.Z0: 0.05, FV.H: 100.0},
-        profiles={FV.WS: "ABLLogWsProfile"},
+        output_vars=[FV.WS, FV.WD, FV.TI, FV.RHO],
+        var2col={},
+        heights=[50, 75, 90, 100, 150, 200, 250, 500],
+        fixed_vars={FV.TI: 0.05},
     )
 
     farm = foxes.WindFarm()
     foxes.input.farm_layout.add_from_file(
-        farm,
-        args.layout,
-        col_x="x",
-        col_y="y",
-        col_H="H",
-        turbine_models=args.tmodels + [ttype.name],
+        farm, args.layout, turbine_models=args.tmodels + [ttype.name],
+        H=170.
     )
+
+    if args.show_layout:
+        ax = foxes.output.FarmLayoutOutput(farm).get_figure()
+        plt.show()
+        plt.close(ax.get_figure())
 
     algo = foxes.algorithms.Downwind(
         mbook,
@@ -45,21 +47,21 @@ def run_foxes(args):
 
     time0 = time.time()
 
-    farm_results = algo.calc_farm()
+    farm_results = algo.calc_farm(vars_to_amb=[FV.REWS, FV.P])
 
     time1 = time.time()
     print("\nCalc time =", time1 - time0, "\n")
 
-    print(farm_results)
+    print(farm_results, "\n")
 
     fr = farm_results.to_dataframe()
-    print(fr[[FV.WD, FV.H, FV.AMB_REWS, FV.REWS, FV.AMB_P, FV.P]])
+    print(fr[[FV.WD, FV.AMB_REWS, FV.REWS, FV.AMB_P, FV.P]])
 
     o = foxes.output.FarmResultsEval(farm_results)
     P0 = o.calc_mean_farm_power(ambient=True)
     P = o.calc_mean_farm_power()
     print(f"\nFarm power: {P/1000:.1f} MW, Efficiency = {P/P0*100:.2f} %")
-
+    
 
 if __name__ == "__main__":
 
@@ -73,8 +75,8 @@ if __name__ == "__main__":
     parser.add_argument(
         "-s",
         "--states",
-        help="The states input file (path or static)",
-        default="states.csv.gz",
+        help="The timeseries input file (path or static)",
+        default="WRF-Timeseries-4464.csv.gz",
     )
     parser.add_argument(
         "-t",
@@ -82,10 +84,14 @@ if __name__ == "__main__":
         help="The P-ct-curve csv file (path or static)",
         default="NREL-5MW-D126-H90.csv",
     )
-    parser.add_argument("-r", "--rotor", help="The rotor model", default="centre")
+    parser.add_argument("-r", "--rotor", help="The rotor model", default="grid9")
     parser.add_argument(
         "-p", "--pwakes", help="The partial wakes model", default="rotor_points"
     )
+    parser.add_argument(
+        "-c", "--chunksize", help="The maximal chunk size", type=int, default=1000
+    )
+    parser.add_argument("-sc", "--scheduler", help="The scheduler choice", default=None)
     parser.add_argument(
         "-w",
         "--wakes",
@@ -96,10 +102,6 @@ if __name__ == "__main__":
     parser.add_argument(
         "-m", "--tmodels", help="The turbine models", default=[], nargs="+"
     )
-    parser.add_argument(
-        "-c", "--chunksize", help="The maximal chunk size", type=int, default=1000
-    )
-    parser.add_argument("-sc", "--scheduler", help="The scheduler choice", default=None)
     parser.add_argument(
         "-n",
         "--n_workers",
@@ -113,6 +115,12 @@ if __name__ == "__main__":
         help="The number of threads per worker for distributed run",
         type=int,
         default=None,
+    )
+    parser.add_argument(
+        "-sl",
+        "--show_layout",
+        help="Flag for showing layout figure",
+        action="store_true",
     )
     parser.add_argument(
         "--nodask", help="Use numpy arrays instead of dask arrays", action="store_true"