+++ foxes/input/states/__init__.py
@@ -1,7 +1,7 @@
 from .single import SingleStateStates
 from .scan_ws import ScanWS
-from .states_table import StatesTable
-from .timeseries import Timeseries
+from .states_table import StatesTable, Timeseries
 from .field_data_nc import FieldDataNC
+from .multi_height import MultiHeightStates, MultiHeightTimeseries
 
 from . import create