+++ docs/source/history.md
@@ -69,3 +69,16 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
   - New notebook: `layout_opt.ipynb`, demonstrating a simple layout optimization case
 
 **Full Changelog**: [](https://github.com/FraunhoferIWES/foxes/commits/v0.2.0)
+
+## v0.2.1-alpha
+
+- Input:
+  - New input: `MultiHeightStates` and `MultiHeightTimeseries`, for input data at several height levels, e.g. WRF results at one point
+- Output:
+  - New output: `FarmResultsEval`, calculates sum, mean, min, max over states or turbines for the whole wind farm
+  - New output: `ResultsWriter`, writes farm results or pandas data to csv file
+  - Renaming: `AmbientRosePlotOutput` is now called `StatesRosePlotOutput`
+- Notebooks:
+  - New notebook: `multi_height.ipynb`, demonstrating the usage of multi-height wind input data
+
+**Full Changelog**: [](https://github.com/FraunhoferIWES/foxes/commits/v0.2.1)